fix: remove max age
