+++ revocation/crl/bundle.go
@@ -1,0 +1,38 @@
+// Copyright The Notary Project Authors.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package crl
+
+import (
+	"crypto/x509"
+	"errors"
+)
+
+// Bundle is in memory representation of the Bundle tarball, including base CRL
+// file and metadata file, which may be cached in the file system or other
+// storage
+//
+// TODO: consider adding DeltaCRL field in the future
+type Bundle struct {
+	// BaseCRL is the parsed base CRL
+	BaseCRL *x509.RevocationList
+}
+
+// Validate checks if the bundle is valid
+func (b *Bundle) Validate() error {
+	if b.BaseCRL == nil {
+		return errors.New("base CRL is missing")
+	}
+
+	return nil
+}
